+++ wolfc/platform.all/AST/WCMulExpr.cpp
@@ -157,7 +157,7 @@ llvm::Constant * MulExprTwoOps::codegenExprConstEval(CodegenCtx & cgCtx) {
 bool MulExprTwoOps::compileCheckBothExprsAreInt() const {
     const DataType & leftType = mLeftExpr.dataType();
     
-    if (!leftType.equals(PrimitiveDataTypes::getUsingTypeId(DataTypeId::kInt))) {
+    if (!leftType.equals(PrimitiveDataTypes::getUsingTypeId(DataTypeId::kInt64))) {
         compileError("Left type in expression must be 'int' for now and not '%s'!",
                      leftType.name().c_str());
         
@@ -166,7 +166,7 @@ bool MulExprTwoOps::compileCheckBothExprsAreInt() const {
     
     const DataType & rightType = mRightExpr.dataType();
     
-    if (!rightType.equals(PrimitiveDataTypes::getUsingTypeId(DataTypeId::kInt))) {
+    if (!rightType.equals(PrimitiveDataTypes::getUsingTypeId(DataTypeId::kInt64))) {
         compileError("Right type in expression must be 'int' for now and not '%s'!",
                      rightType.name().c_str());
